refactor: style chrome Storage
