+++ src/hooks/useTheme.ts
@@ -1,28 +1,9 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { useEffect } from 'react';
-import { useColorScheme, Appearance } from 'react-native';
+import { useColorScheme } from 'react-native';
 
-import { getPlayerSkin, savePlayerSkin } from '../utils/ChromeStorage';
-import { createStyle } from '../components/style';
-import { resolveBackgroundImage } from '../components/background/MainBackground';
+import { getPlayerSkin } from '../utils/ChromeStorage';
 import { useNoxSetting } from '@stores/useApp';
-
-export const savePlayerStyle = async (
-  val: NoxTheme.Style | NoxTheme.AdaptiveStyle,
-  save = true
-) => {
-  const createFromStyle = val.isAdaptive
-    ? Appearance.getColorScheme() === 'dark'
-      ? val.darkTheme
-      : val
-    : val;
-  const createdStyle = createStyle(createFromStyle);
-  if (save) savePlayerSkin(val);
-  return {
-    ...createdStyle,
-    bkgrdImg: await resolveBackgroundImage(createdStyle.bkgrdImg),
-  };
-};
 
 const useTheme = () => {
   const setPlayerStyleState = useNoxSetting(state => state.setPlayerStyle);
