+++ src/utils/StyleStorage.ts
@@ -1,0 +1,22 @@
+import { Appearance } from 'react-native';
+
+import { savePlayerSkin } from './ChromeStorage';
+import { createStyle } from '../components/style';
+import { resolveBackgroundImage } from '../components/background/MainBackground';
+
+export const savePlayerStyle = async (
+  val: NoxTheme.Style | NoxTheme.AdaptiveStyle,
+  save = true
+) => {
+  const createFromStyle = val.isAdaptive
+    ? Appearance.getColorScheme() === 'dark'
+      ? val.darkTheme
+      : val
+    : val;
+  const createdStyle = createStyle(createFromStyle);
+  if (save) savePlayerSkin(val);
+  return {
+    ...createdStyle,
+    bkgrdImg: await resolveBackgroundImage(createdStyle.bkgrdImg),
+  };
+};
